Update test files for UInt24 type changes
Cast int literals to UInt24 in SetUInt24/WriteUInt24/AddUInt24 calls.
Cast UInt24 results to int before Should().Equal() assertions.
Fix overflow value 0x78654321 to valid UInt24 value 0x654321.

+++ src/MrKWatkins.BinaryPrimitives.Tests/ByteReadOnlySpanExtensionsTests.cs
@@ -64,7 +64,7 @@ public sealed class ByteReadOnlySpanExtensionsTests
     {
         ReadOnlySpan<byte> bytes = [0x78, 0x56, 0x34];
 
-        bytes.GetUInt24().Should().Equal(0x345678);
+        ((int)bytes.GetUInt24()).Should().Equal(0x345678);
     }
 
 
@@ -73,8 +73,8 @@ public sealed class ByteReadOnlySpanExtensionsTests
     {
         ReadOnlySpan<byte> bytes = [0x78, 0x56, 0x34];
 
-        bytes.GetUInt24(Endian.Little).Should().Equal(0x345678);
-        bytes.GetUInt24(Endian.Big).Should().Equal(0x785634);
+        ((int)bytes.GetUInt24(Endian.Little)).Should().Equal(0x345678);
+        ((int)bytes.GetUInt24(Endian.Big)).Should().Equal(0x785634);
     }
