+++ src/MrKWatkins.BinaryPrimitives.Tests/EndianExtensionsTests.cs
@@ -5,7 +5,7 @@ public sealed class EndianExtensionsTests
     [TestCase(Endian.Big, 0x12, 0x34, 0x56, 0x00123456)]
     [TestCase(Endian.Little, 0x12, 0x34, 0x56, 0x00563412)]
     public void ToUInt24(Endian endian, byte byte0, byte byte1, byte byte2, int expected) =>
-        endian.ToUInt24(byte0, byte1, byte2).Should().Equal(expected);
+        ((int)endian.ToUInt24(byte0, byte1, byte2)).Should().Equal(expected);
 
 
     [Test]
